requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.6.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
redis==5.0.1
schedule==1.2.0
newspaper3k==0.2.8
requests==2.31.0
pandas==2.1.4
jinja2==3.1.2
websockets==12.0